检查文件存在再clear

+++ clear.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
 
-rm SourceHanSansHWSC.zip
-rm -r Open_Data/
+if [ -e "SourceHanSansHWSC.zip"] ; then
+    rm SourceHanSansHWSC.zip  
+fi
+if [ -d "Open_Data"] ; then
+    rm -r Open_Data
+fi
